server_5_analytics/requirements/dev.txt
fastapi==0.115.0
pydantic==2.9.2
python-dotenv==1.0.1
uvicorn==0.30.6
requests==2.32.3
pandas==2.2.2
redis==5.0.8
prophet==1.1.5
torch==2.4.1
numpy<2.0
